requirements.txt
opencv-python==4.8.0
Pillow==10.0.0
pandas==2.0.3
qrcode==7.4.2
pyzbar==0.1.9
num2words==0.5.12
numpy==1.24.3